:recycle: Move Flask-Login to an initializer.
